bib_postgre_func/pyproject.toml
[tool.poetry]
name = "postgre_func"
version = "0.1.0"
description = ""
authors = ["Kaua Hen, Emilly, Jonas, Wallison, Flavio"] 
readme = "README.md"
packages = [
    { include = "postgre_func", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "2.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
